moving mount functions to utils
