+++ perfmetrics/scripts/hns_rename_folders_metrics/generate_folders_and_files.py
@@ -22,7 +22,6 @@ import logging
 import os
 import sys
 import subprocess
-from subprocess import Popen
 
 OUTPUT_FILE = str(dt.now().isoformat()) + '.out'
 TEMPORARY_DIRECTORY = './tmp/data_gen'
