+++ perfmetrics/scripts/hns_rename_folders_metrics/generate_folders_and_files_test.py
@@ -527,109 +527,5 @@ class TestParseAndGenerateDirStructure(unittest.TestCase):
     mock_subprocess.assert_has_calls(expected_subprocess_calls)
 
 
-class TestParseAndGenerateDirStructure(unittest.TestCase):
-
-  @patch('generate_folders_and_files.LOG_ERROR','error')
-  @patch('generate_folders_and_files._logmessage')
-  def test_no_dir_structure_passed(self,mock_logmessage):
-    dir_str={}
-
-    exit_code=generate_folders_and_files._parse_and_generate_directory_structure(dir_str)
-
-    self.assertEqual(exit_code,1)
-    mock_logmessage.assert_called_once_with("Directory structure not specified via config file.",'error')
-
-
-  @patch('generate_folders_and_files.TEMPORARY_DIRECTORY', './tmp/data_gen')
-  @patch('subprocess.call')
-  @patch('generate_folders_and_files._logmessage')
-  @patch('generate_folders_and_files._generate_files_and_upload_to_gcs_bucket')
-  def test_valid_dir_str_with_folders(self, mock_generate, mock_log, mock_subprocess):
-    dir_str = {
-        "name": "test_bucket",
-        "folders": {
-            "num_folders":1,
-            "folder_structure": [
-                {
-                    "name": "test_folder",
-                    "num_files": 2,
-                    "file_name_prefix": "file",
-                    "file_size": "1kb"
-                }
-            ]
-        },
-        "nested_folders": {
-            "folder_name": "test_nested",
-            "num_folders": 1,
-            "folder_structure": [
-                {
-                    "name": "test_nested_folder1",
-                    "num_files": 2,
-                    "file_name_prefix": "file",
-                    "file_size": "1kb"
-                }
-            ]
-        }
-    }
-    mock_generate.return_value= 0
-    expected_subprocess_calls = [
-        call(['mkdir', '-p', generate_folders_and_files.TEMPORARY_DIRECTORY]),
-        call(['rm', '-r', generate_folders_and_files.TEMPORARY_DIRECTORY])
-    ]
-    expected_generate_and_upload_calls = [
-        call( 'gs://test_bucket/test_folder/',2,'kb',1,'file'),
-        call('gs://test_bucket/test_nested/test_nested_folder1/',2,'kb',1,'file')
-    ]
-
-    exit_code = generate_folders_and_files._parse_and_generate_directory_structure(dir_str)
-
-    self.assertEqual(exit_code, 0)
-    # Verify subprocess calls.
-    mock_subprocess.assert_has_calls(expected_subprocess_calls)
-    # Verify log messages.
-    mock_log.assert_any_call('Making a temporary directory.\n', 'info')
-    mock_log.assert_any_call('Deleting the temporary directory.\n', 'info')
-    # Verify generate_files_and_upload_to_gcs_bucket call.
-    mock_generate.assert_has_calls(expected_generate_and_upload_calls)
-
-
-  @patch('generate_folders_and_files.TEMPORARY_DIRECTORY', './tmp/data_gen')
-  @patch('subprocess.call')
-  @patch('generate_folders_and_files._logmessage')
-  @patch('generate_folders_and_files._generate_files_and_upload_to_gcs_bucket')
-  def test_create_folder_failure(self, mock_generate, mock_log, mock_subprocess):
-    dir_str = {
-        "name": "test_bucket",
-        "folders": {
-            "num_folders":1,
-            "folder_structure": [
-                {
-                    "name": "test_folder",
-                    "num_files": 2,
-                    "file_name_prefix": "file",
-                    "file_size": "1kb"
-                }
-            ]
-        }
-    }
-    mock_generate.return_value= 1
-    expected_subprocess_calls = [
-        call(['mkdir', '-p', generate_folders_and_files.TEMPORARY_DIRECTORY])
-    ]
-    expected_generate_and_upload_calls = [
-        call( 'gs://test_bucket/test_folder/',2,'kb',1,'file')
-    ]
-
-    exit_code = generate_folders_and_files._parse_and_generate_directory_structure(dir_str)
-
-    self.assertEqual(exit_code, 1)
-    # Verify log messages.
-    mock_log.assert_any_call('Making a temporary directory.\n', 'info')
-    # Verify generate_files_and_upload_to_gcs_bucket call.
-    mock_generate.assert_has_calls(expected_generate_and_upload_calls)
-    # Verify subprocess calls.
-    mock_subprocess.assert_has_calls(expected_subprocess_calls)
-
-
 if __name__ == '__main__':
   unittest.main()
